+++ Common/Subworlds/MappingWorld.cs
@@ -1,7 +1,9 @@
 ﻿using System.Collections.Generic;
 using PathOfTerraria.Common.Subworlds.Passes;
 using PathOfTerraria.Common.Systems.DisableBuilding;
+using ReLogic.Graphics;
 using SubworldLibrary;
+using Terraria.GameContent;
 using Terraria.IO;
 using Terraria.WorldBuilding;
 
@@ -48,5 +50,27 @@ public abstract class MappingWorld : Subworld
 		WorldGenerator.CurrentGenerationProgress = progress;
 		Main.ActiveWorldFileData.SetSeedToRandom();
 		GenVars.structures = new();
+	}
+
+	public override void DrawMenu(GameTime gameTime)
+	{
+		string statusText = Main.statusText;
+		GenerationProgress progress = WorldGenerator.CurrentGenerationProgress;
+
+		if (WorldGen.gen && progress is not null)
+		{
+			DrawStringCentered(progress.Message, Color.LightGray, new Vector2(0, 60), 0.6f);
+			double percentage = progress.Value / progress.CurrentPassWeight * 100f;
+			DrawStringCentered($"{percentage:#0.##}%", Color.LightGray, new Vector2(0, 120), 0.7f);
+		}
+
+		DrawStringCentered(statusText, Color.White);
+	}
+
+	private static void DrawStringCentered(string statusText, Color color, Vector2 position = default, float scale = 1f)
+	{
+		Vector2 screenCenter = new Vector2(Main.screenWidth, Main.screenHeight) / 2f + position;
+		Vector2 halfSize = FontAssets.DeathText.Value.MeasureString(statusText) / 2f * scale;
+		Main.spriteBatch.DrawString(FontAssets.DeathText.Value, statusText, screenCenter - halfSize, color, 0f, Vector2.Zero, scale, SpriteEffects.None, 0);
 	}
 }
